refactor: split API handler responsibilities
