+++ architecture.md
@@ -1068,6 +1068,8 @@ The CLI has a two-layer architecture: a Node.js wrapper (`packages/rslint/src/cl
 
 The default Go CLI mode keeps one lint orchestration path in `lint_pipeline.go`. Supporting files own bounded command-layer concerns: `lint_options.go` parses invocation arguments and validates flag combinations; `lint_configuration.go` contains configuration and rule-catalog preparation shared by CLI and API; `lint_config_resolver.go` resolves the effective configuration for each selected file; `lint_diagnostics.go` normalizes diagnostics and explicit-file warnings; `lint_fix.go` owns one fix application and disk-write pass, while the pipeline retains multi-pass rebuild orchestration; `eslint_plugin.go` adapts prepared lint plans to third-party plugin dispatch; and `syntax_diagnostics.go` collects target syntax diagnostics. Target selection remains in `internal/config/target`, Program construction remains in `internal/program/loader`, and report rendering remains in `internal/output`.
 
+The Go API mode keeps one lint-request orchestration path in `api_lint.go`. `api_handler.go` binds `IPCHandler` to the IPC service and reverse plugin transport; `api_config_module_loader.go` adapts reverse config loading and activation; `api_lint_filesystem.go` owns request-scoped overlay and canonical-path projection; and `api_ast_info.go` owns the separate inspector request and its content/options-keyed Program cache. The API-specific path, structured-response, and single-pass in-memory fix policies remain at this command boundary, while config discovery, target planning, Program loading, prepared lint plans, native/plugin execution, and the pure fix algorithm remain shared with the CLI through their existing packages.
+
 1. **Node.js Wrapper**: parses args, starts the Go engine, and hosts JS/TS module evaluation plus live third-party plugin objects
 2. **Config Catalog**: for automatic JS/TS discovery or an explicitly selected JS/TS config, Go builds the staged catalog and batches exact module-evaluation requests to Node. If automatic discovery finds no candidate, or a non-JS config was explicitly selected, the existing Go JSON loader path remains in control
 3. **Mode Selection**:
